+++ tests/Feature/FileUploadTest.php
@@ -18,25 +18,34 @@ class FileUploadTest extends TestCase
    *
    * @return void
    */
-  public function testIndex()
+  public function test_index()
   {
     $response = $this->get('/upload');
 
     $response->assertStatus(200);
   }
 
-  public function testFileValidation()
+  public function test_file_validation_with_no_data()
   {
-    $response = $this->post('upload');
+    $response = $this->json('POST', 'upload',[]);
+    $response->assertStatus(422);
+  }
+
+  public function test_file_validation_with_wrong_no_data()
+  {
+    $data = [
+      'file' => UploadedFile::fake()->image('logo.png')
+    ];
+    $response = $this->json('POST', 'upload',$data);
     $response->assertStatus(422);
   }
 
   public function testCreateSuccessfully()
   {
     Bus::fake();
-    $file = resource_path('test/test.csv');
+    $file = resource_path('test/test.json');
     $data = ['file' => $file];
-    $response = $this->post('/upload', $data);
+    $response = $this->json('POST','upload', $data);
     Bus::assertDispatched(SaveUserDetails::class);
     $response->assertOk();
   }
